Gen 1 Tool for checking and creating version control.
The python function to create a markdown file that holds all the updates that are being specified in the release notes meathod.
This is a very rough tool and should be worked on and improved around the summary system. I need to look for or make a better summary tool. As the current tool I am using is sometimes just not good.