This is a revised version of the tests that have been generated by AI.
The next step for this kind of test is ensure that is is fitting into the norms that were set out over the summer. I intend on doing this by hand for now until I update the model's code. This is to ensure that if there is a failure in the code it will be the generated code not the implementation in the test suit.